Add env var to determine if build was skipped.
This change is being made to try and fix broken storybook deployment on master.

Previously it was not able to deploy storybook if the build was not performed because the static assets had not been build. With this change a new condition has been added that should prevent deployments to GitHub pages if no build was performed.