Bug - Piping to the front after changes
The cursor position was being lost when a new value came back into
the rich text editor after a change. The change was being saved as
on selecting to pipe the rich text editor is blurred so entityEditor
is triggered and the change is sent to the server. When this comes
back to the RTE it was losing the selection state from the draft js
editor state.

This change forces selection to the previous value before the update.

There is still an issue where cursor location is set back to 0 on
cancelling piping.