Check JWT token on graphql API requests. This change stores an access token generated by Firebase in localStorage. This access token is then passed to the graphql API via the Apollo HTTP link headers.
The API checks for a valid token and places the decoded token onto the GraphQL context. In the event that the token is invalid or not present, then the API will respond with a 401 Unauthorized status unless the API is running in development mode.

A new query has been added to the GraphQL API to retrieve details about the current user. This can be accessed via the `me` query. This should be called by author in order to display and retrieve details about the identity of the currently logged in user. This token based authentication forms the basis of the future permissions and security model of Author.

Now that the API declines any unauthorized request with a 401 response, a change is required to publisher to generate and send a valid auth token to the API with each query it issues on launch.

I have temporarily implemented this logic as a series of middleware functions. This will likely be replaced soon once the publishing logic is merged into the API