Bug - Unable to pipe dates
This was caused by us filtering out the date answers from the answers
you could pipe.

This change also fixes an issue where the get previous answers would
be returned in a non-deterministic order. So now we order in the db so
we can guarantee the order of the answers.