Debounce was casuing data to be lost
When a field is changed then the dirtyField property is set. On
blur of the field update is called where dirtyField is cleared as the
request to update is triggered.

When update was debounced and hence delayed by 20ms then the props to
which entity the dirtyField was for and which request to update would
be triggerd could change between the dirtyField being set and the actual
update running.

This occurred when navigating to the same type of entity and blurring a
field at the same time.

Remove frequently failing cypress tests that debounce tried to fix. We
will resolve section introduction issue later.