Fix race condtion in toggling/updating validation
Sometimes when changing an input and then toggling a validation to be
enabled/disabled then a race condition could ocurr causing apollo cache
to not be correct. We can resolve this by only fetching the necessary
fields for each mutation so that they do not clash and possibly update
themselves in the wrong order.

This possibly fixes #157