Context shared across requests
We were sharing the context & therefore questionnaire across requests
which was causing dirty reads.

For section intro there was a race condition where the mutation for
createSectionIntro was running interlaced with updateSection.

There are 3 stages to each request:

1. the read of the questionnaire from the db (in middleware)
2. the mutation - running the mutation resolver and triggering the write
3. the read of the values on return (using the return value in
the resolvers)

They are broken up like this due as between these are the async points
where other things that were queued up can start in-between and run
until their async point.

The problem we had was that all parts were relying on
`ctx.questionnaire` which was a static object and so `ctx.questionnaire`
could change when requests happened at the same time.