Fix issue where if branch is not master no tag was empty.
There was a bug in the existing logic where the tag used when pushing up docker images would be empty if a branch other than master was built.
This change sets the tag to be the branch name if the branch being built is not master and it is not a pull request.

Exit the build if the developer pushes up a branch called latest.

Co-Authored-By: samiwel <samiwel@users.noreply.github.com>