Fix flakey duplicate questionnaire spec
Sometimes the afterEach of deleting a questionnaire would fail. It would
report that the delete button was disabled. This was ocurring only in
the duplicate questionnaire spec.

This was because the optimistic update in the UI shows a row and so the
assertion was passing but the server was not actually finshed. This
meant the after each was trying to delete an optimistic response which
is not possible.

Now, to resolve this instead we wait for the link to be enabled which
means the server has responed. This means it can then be deleted by the
afterEach.