Stop crashing on invalid questionnaire id
This fix changes it so when a questionnaire cannot found that the
`ctx.questionnaire` is null in the resolvers. This means that we see
a graphql error in the UI.

Before this change it was blowing up in the express middleware and so
the server was crashing and the requests were left pending.