Make the versions a shadow table
Change the way we use the versions table so we the latest version of any
questionnaire on the questionnaires table and all versions of a
questionnaire on the versions table.

This means we read and save all data from the questionnaire table
(no title bug) and we no longer have to perform a scan to get the
latest questionnaire. We can just look it up by id.