Fix piping not decorating currency answers.
There was a defect where the £ symbol was not being added to piped currency answers.
I tracked down the cause being that the type was not being added to the piping data attributes.

The type is required at present and used by publisher in order to determine how to correctly render the piped output in runner.
A change was made to shapeTree to ensure that the type property is present on answers when selecting values to pipe.