Fix issue where error thrown when creating a routing
1. API was incorrectly creating a right side for a new binary expression
when it should be null. The right side it was creating wasn't valid for
all scenarios either - it was fine for Radio but not basic answers.

2. Input was switching from uncontrolled to controlled causing React to
log an error which caused Cypress to blow up. This was caused by the
right side value defaulting to undefined not found. Have change it to
null which resolves this.