Questionnaire introduction
Introduce new questionnaire introduction for all business surveys.
Add introduction and default metadata for existing business surveys.

Notable changes
1. Create Reorder component for re-ordering answers and collapsibles.
2. Add test to ensure that migrations are deterministic.
3. After discovering an issue with jsondiffpatch where array items where
being added when a patch was applied when they already existed. We have
to check if the patch needs to be applied before applying it when
dealing with conflicts in dynamo datastore.