Reduce what we return in a scan
The scan was very slow as it was returning all the data in author. So
now instead, we store just the data needed for the listing page in the
list table which will still scan.

Then when we are reading we read all the data from the versions table
and ensure we are reading the latest version by the sort/range key on
the field updatedAt.

This takes a scan of ~160 questionnaires from >2 seconds to 170ms.