Add questionnaire table pagination
This has added a reducer for managing the interactions on the
questionnaire list table to make adding sorting and filtering easier
later on.

It also use react-testing-library to perform behaviour integration tests
across the set of components used. This made it easier to test a
write the tests in a black box way so we can change the internal
implementations much easier without losing anything.