Attempt to fix flakey calculated summary spec
This was failing when going to preview and the total title was not set.
I managed to recreate this locally once and it was due to the answer
shortcut click response containing a blank title. It must have been out
of sync.

So this now waits for the total title to have finished saving by
checking the lack of saving indicator.