Fix issue where unable to delete questionnaires
After restricting deletion to only users with write access we now
require having the questionnaire in the api request context to determine
if a user can delete a questionnaire.

This context is populated by middleware reading the questionnaireId from
the headers. For delete questionnaire this was not being sent as it is
on a list page and client middleware generating the header from the URL
for all other mutations does not work here so we need to send it
manually.