Fix flashing side bar when adding question page
Depending on how you load your cache you can create a situation where
an error is thrown when you add a question page. This causes a
noticeable flash to the user.

This is because apollo does not seem to be able to merge structures on
an object unless that is an object with its own id and normalized in the
apollo cache.

This change now returns the validation error info with an id so it can
be normalized. This then means that the cache is merged correctly.