Simplify setup of GTM.
The way in which we were setting up GTM was more complex than it needed to be.

There is an existing library which integrates nicely with React applications which takes care of injecting the relevant GTM snippets.

We were also hard coding the preview Id which made it impossible to successfully turn on GTM in both pre-prod and production environments.

This change renames the existing ENV_ID environment variable to GTM_AUTH and introduces a separate GTM_PREVIEW environment variable to simplify the configuration across environments and bring the terminology inline with what GTM calls these variables.