Min/Max val text reads 'must be (higher/lower) than' when given val is included
Link to Trello story: https://trello.com/c/VCD3FsH7/1353-min-max-text-on-numerical-values-3

Modified AnswerValidation to make it so that the 'Max value' and 'Min value' text shows 'must be higher than' and 'must be lower than', respectively, when the given value is specified to be included. When it is not included, the word 'value' is now lowercase (previous uppercase). Tests made and are passing when evaluating every number type (number, percentage, etc.) for both min and max value phrases.