♻️ Allow fromHTML to process multiple entities
Updates fromHTML to be in the style of the existing toHTML function - it
now accepts a map of node names to functions rather than only accepting
one entity's function.

This is needed in order to allow more entities' conversion code to be
added in a modular way (e.g. to add the link conversion code when
written).