🛠♻️ Add Text_Optional in gql enum; refactor consts
Adds support for the Text_Optional type in author-api's enum of possible
metadata types.

Re-factor existing defaultMetadata code to use the available constants
instead of hardcoded strings

Re-factor the author front end metadata constants file to decouple the
visible labels of metadata from their graphql enum representation; this
was needed so that a space and parenthesis can be present in the
"Text (Optional)" select label.