✅ Fix Author API tests
The firestore test doesn't seem to be directly related to code changes
I've made - but I've updated it so the syntax is right for checking what
the actual code does. The function it references throws an exception if
no ID is found so I've adjusted the test to make that the desired
behaviour.