🛠 Fix region / language metadata dropdown
Previous PR de-coupled the labels of the select options in the metadata
editor from the values (as recognised by the backend server). Forgot to
make this change for the "region" and "language" constant types however
which meant that author UI broke when these were chosen as metadata
types.

This PR updates these types to be consistent with the new approach.