✨ When only one expression, clear group operator
Routing only - implemented clientside so as to not interfere with skip
conditions.

New optional callback added to BinaryExpressionEditor which allows
specifying code to run after a successful expression deletion has taken
place.

This is used by the Routing RuleEditor to clear group operator ("Select
AND/OR") per AC when penultimate expression deleted.