✨ EAR 349 - Add ability to star questionnaires
Data re: which questionnaires are starred is kept on the users db table,
as the array `starredQuestionnaires`.

The UI code doesn't need to be aware of this, however - there's a
`starred` resolver for the Questionnaire object which will return
whether the currently logged in user has starred the questionnaire in question.