EAR 1207 - ✨ Add validation error when all themes disabled (#2032)
* ✨ API: Validate when all themes disabled

* ♻️ Themes: refactor API design

- Use `ThemeSettings` wrapper structure to encapsulate theme-related
attributes (`themes`, `previewTheme`) and provide a basis for
`validationErrorInfo` resolution for errors not associated with any one
particular theme.
- Make `Theme` `id` always equal to `shortName` - this means the API can
always return a list of known themes, even if not present in the JSON,
to the client. The client then does not need to manually modify the
cache or be aware of the actual enumeration of possible themes. Client
code for ThemesPage simplified significantly after this change.
- Add migrations to new structures & to make `shortName` equivalent to `id`
  for existing `Theme` instances.