Fix to make the search case insensitive
There’s a bug here in the search, where we’re matching in the filter function against `displayName`. In the dummy data that we’re passing in though, the `displayName` for the first few questions is `1a` and `1b` so they don’t match.

I don’t know if that’s expected behaviour or if we should be matching it against the `title` in the data and not `displayName`, or if the mock data is the wrong shape.