EAR-1414-move-delete-options-not-visible-on-answer-options (#2171)
* Fixes to the button

Fixes to the button not being shown. Pretty sure the colour has been switched to white.

* New `color` prop

Added a new color prop to the icons. Defaults to the secondary colour as that seems to be the default through most of the app. Passing `white` will conditonally change the colour.

* Answer Editor Colour

Changed the answer editor colour to `white` to match the current design.

* Removed a duplicate key causing merge to fail

* Jest Snapshot Changes