Ear 1317 modal for importing questions (#2161)
* Setup

* Work

* Added search functionality

* Added picker v2

* Searching works

* Revert QuestionPicker; added mock data

* Moved things around

* Renamed

* Fixed tests

* Linting

* Linting

* Snapshots

* Added a test

* Removed desribe.only

* Testing Files

First round of testing for the modal questionnaire picker.

* Styled Component Bug

Getting a weird error where the styled component was trying to effect things in the DOM. Looking into it, the solutions are things that have already been done in the file. One workaround, was to use a template literal on the prop.

* Updates to tests

Updates to tests.

* Updates to tests

Updated tests so that there is a `not` condition for searches so prevent any false positives.

* Fix to make the search case insensitive

There’s a bug here in the search, where we’re matching in the filter function against `displayName`. In the dummy data that we’re passing in though, the `displayName` for the first few questions is `1a` and `1b` so they don’t match.

I don’t know if that’s expected behaviour or if we should be matching it against the `title` in the data and not `displayName`, or if the mock data is the wrong shape.

* Removed unecessary code

* Update eq-author/src/components/QuestionPicker/index.test.js

Co-authored-by: tbaggaley <60441612+tbaggaley@users.noreply.github.com>

* Fix for the search bug

* Update eq-author/src/components/QuestionPicker/index.js

Co-authored-by: tbaggaley <60441612+tbaggaley@users.noreply.github.com>

* Update eq-author/src/components/QuestionPicker/index.js

Co-authored-by: tbaggaley <60441612+tbaggaley@users.noreply.github.com>

* Update eq-author/src/storybook/Patterns/Importing content/SelectQuestions.stories.js

Co-authored-by: tbaggaley <60441612+tbaggaley@users.noreply.github.com>

* Update eq-author/src/components/QuestionPicker/index.js

Co-authored-by: tbaggaley <60441612+tbaggaley@users.noreply.github.com>

* Tests and linting

* Update eq-author/src/storybook/Patterns/Importing content/SelectQuestionnaire.stories.js

Co-authored-by: farres1 <71004932+farres1@users.noreply.github.com>

* Update eq-author/src/storybook/Patterns/Importing content/SelectQuestions.stories.js

Co-authored-by: farres1 <71004932+farres1@users.noreply.github.com>

Co-authored-by: Ashley Tucker <ashley.tucker@ons.gov.uk>
Co-authored-by: Ashley Tucker <ashley.rhys.tucker@gmail.com>
Co-authored-by: tbaggaley <60441612+tbaggaley@users.noreply.github.com>
Co-authored-by: farres1 <71004932+farres1@users.noreply.github.com>